requirements.txt
asgiref==3.8.1
blinker==1.9.0
channels==4.2.2
click==8.2.0
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==1.16.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mongoengine==0.27.0
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pymongo==4.6.0
pytesseract==0.3.13
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
